Completely overhauls rubrics. Make sure you have the correct supporting files that were just sent out in an email.

+++ trunk/src/config/LabPart.java
@@ -16,7 +16,7 @@ public class LabPart extends Part
     }
 
     //TODO: Implement parsing of lab score / completion
-    public int getScore(String studentLogin)
+    public double getScore(String studentLogin)
     {
         return 0;
     }
